Reorganized namespaces
Moved the following into TDkit.Chemistry
- ChemicalFormula.cs
- Element.cs
- Fragment.cs
- Isotope.cs
- Modification.cs
- Polymer.cs
- Proteoform.cs
- Residue.cs
