+++ TDkit/MassSpec/Mercury7.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
+using TDkit.Chemistry;
 
 namespace TDkit.MassSpec
 {
