+++ TDkit/MassSpec/Transition.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using TDkit.Chemistry;
 
 namespace TDkit.MassSpec
 {
