+++ TDkitTest/Chemistry/FragmentTest.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using TDkit;
+using TDkit.Chemistry;
 
-namespace TDkitTest
+namespace TDkitTest.Chemistry
 {
     [TestClass]
     public class FragmentTest
